requirements.txt
openai
supabase
python-dotenv
replicate
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi
uvicorn
